Allow load server config from file and env vars
Use mehcode/config crate to configure the sources and priorites for
configuration loader.

I've moved all structs on `server/mod.rs` to a config module and I've
splitted in two distinct structures:

- RawConfig (I couldn't find a better name) which holds the
  configuration that can be changed by the users (via a TOML file or env
vars)
- Config: The calculated and parsed version, which holds all the
  materialized configuration.

As `CONFIG` is a static value and is used as a singleton on some places,
I've needed to add the `MAELSTROM_CONFIG_FILE` env var to allow loading
from a distinct file on tests.

I wanted to ask some questions:
- Are any concerns on depending on `mehcode/config` and/or `anyhow`?
- Which config format is prefered?  TOML? YAML? dotenv?
- Ia there any problem with prefix all env vars with `MAELSTORM`?